sdsdsds

+++ scene/3d/camera_3d.cpp
@@ -794,9 +794,9 @@ void Camera3D::set_cull_mask_value(int p_layer_number, bool p_value) {
 	ERR_FAIL_COND_MSG(p_layer_number > 20, "Render layer number must be between 1 and 20 inclusive.");
 	uint32_t mask = get_cull_mask();
 	if (p_value) {
-		mask |= 1 << (p_layer_number - 1);
+		mask |= 1 << ((uint32_t)p_layer_number - 1U);
 	} else {
-		mask &= ~(1 << (p_layer_number - 1));
+		mask &= ~(1 << ((uint32_t)p_layer_number - 1));
 	}
 	set_cull_mask(mask);
 }
